Support configuration environment variables.
Add support for OAUTH_CALLBACK_URL environment variable.

Allow the OAuth callback/redirect URL to be specified with an environment variable.

The combination of OAUTH_CALLBACK_URL, OAUTH_CLIENT_ID and OAUTH_CLIENT_SECRET as environment variables allow instances of this app to be launched seperately without requiring an code changes & builds.

Add support for a REDIS_HOST & REDIS_PORT environment variable.

For situations where Redis is on a seperate host, but not using the Redis To Go URL mechanisms.

Add support for OAUTH_URL and API_URL environment vars.

Set OAUTH_URL environment to override the URL used for OAuth token generation & exchanges.

Set API_URL to override the URL used for all API calls.

Usually these two are set together (though not to the same value) to refer to a development or testing stack as opposed to the main Onshape production stack.